IMPALA-10851: Codegen for structs
IMPALA-9495 added support for struct types in SELECT lists but only with
codegen turned off. This commit implements codegen for struct types.

To facilitate this, code generation for reading and writing 'AnyVal's
has been refactored. A new class, 'CodegenAnyValReadWriteInfo' is
introduced. This class is an interface between sources and destinations,
one of which is an 'AnyVal' object: sources generate an instance of this
class and destinations take that instance and use it to write the value.

The other side can for example be tuples from which we read (in the case
of 'SlotRef') or tuples we write into (in case of materialisation, see
Tuple::CodegenMaterializeExprs()).  The main advantage is that sources
do not have to know how to write their destinations, only how to read
the values (and vice versa).

Testing:
  - enabling tests for structs in select list in codegen mode in
    tests/query_test/test_nested_types.py

Change-Id: I5272c3f095fd9f07877104ee03c8e43d0c4ec0b6