IMPALA-13247: Support Reading Puffin files for the current snapshot
This change adds support for reading NDV statistics from Puffin files
when they are available for the current snapshot. Puffin files or blobs
that were written for other snapshots than the current one are ignored.

The NDV values read from Puffin files take precedence over NDV values
stored in the HMS. This is because we only read Puffin stats for the
current snapshot, so these values are always up-to-date, while the
values in the HMS may be stale.

Note that it is currently not possible to drop Puffin stats from Impala.
For this reason, this patch also introduces two ways of disabling the
reading of Puffin stats:
  - globally, with the "--disable_reading_puffin_stats" startup flag:
    when it is set to true, Impala will never read Puffin stats
  - for specific tables, by setting the
    "impala.iceberg_disable_reading_puffin_stats" table property to
    true.

Note that this change is only about reading Puffin files, Impala does
not yet support writing them.

Testing:
 - created the PuffinDataGenerator tool which can generate Puffin files
   and metadata.json files for different scenarios (e.g. all stats are
   in the same Puffin file; stats for different columns are in different
   Puffin files; some Puffin files are corrupt etc.). The generated
   files are under the "testdata/ice_puffin/generated" directory.
 - The new test class 'test_iceberg.py::TestIcebergTableWithPuffinStats'
   uses the generated data to test various scenarios.
 - Added custom cluster tests in test_iceberg_with_puffin_stats.py that
   test the 'disable_reading_puffin_stats' startup flag.

Change-Id: I50c1228988960a686d08a9b2942e01e366678866